distilBert/requirements.txt
torch
datasets
transformers
kafka-python